requirements.txt
langchain==0.3.20
langchain-community==0.3.20
chromadb==0.6.5
sentence-transformers==3.4.1
streamlit==1.42.0
pypdf==6.1.0
python-dotenv==1.0.1
